Adds support for iOS.
