+++ package.json
@@ -10,6 +10,7 @@
     "bluebird": "^3.4.6",
     "react": "15.3.2",
     "react-native": "0.34.1",
+    "react-native-drawer-layout": "^1.1.0",
     "react-native-vector-icons": "^2.1.0",
     "react-redux": "^4.4.5",
     "redux": "^3.6.0",
